+++ src/traces/pie/attributes.js
@@ -18,7 +18,7 @@ var extendFlat = require('../../lib/extend').extendFlat;
 var textFontAttrs = fontAttrs({
     editType: 'calc',
     arrayOk: true,
-    colorEditType: 'style',
+    colorEditType: 'plot',
     description: 'Sets the font used for `textinfo`.'
 });